requirements.txt
pick
lmstudio
requests
crawl4ai
python-dotenv
pytest
pytest-asyncio
pytest-mock